\name{openCheatsheet}
\alias{openCheatsheet}
\title{Open Cheat Sheet}
\usage{
openCheatsheet()
}
\description{
This addin let's the user select an R cheat sheet and open its github page in a web browser or, for the list of R functions used in swirl course 'Landscape_Genetics_R_Course'), as an editable .doc file.
}

